requirements.txt
Flask==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
 requests==2.26.0
 tensorflow==2.5.0
opencv-python==4.5.3.56
 autopep8==1.6.0
cffi==1.15.0
click==8.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycodestyle==2.8.0
pycparser==2.21
toml==0.10.2
Werkzeug==2.1.1
zipp==3.8.0
gunicorn==20.1.0
 cryptography==3.4.7
 Flask-Cors==3.1.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-Script==2.0.6
numpy==1.21.2
tensorflow==2.6.0
keras==2.6.0
opencv-python-headless==4.5.3.56